backend/requirements.txt
Flask
flask_sqlalchemy
flask-cors
requests
python-dotenv